IntegerSquareRootD

Description

Documentation

integerSquareRoot

Arguments

:: Int
The number for which to calculate the integer square root.

-> Maybe Int
The integer square root or Nothing, if one cannot be found.

The integerSquareRoot function calculates the integer square root of a given number.

doSquareRoot

Arguments

:: Int
The number for which to calculate the integer square root

-> Int
The current "guess"

-> Maybe Int
The integer square root or Nothing, if one cannot be found.

Finds the integer square root of a given number. Starts by calculating whether the guess (second argument) is an integer square root of the given number (first argument). Calls itself recursively if it is not.